Revert "Add sampler benchmarking code"
This reverts commit 9ae87ccd6010dbf4d346cd6991e86022e64f4e36.
